Fix memory leak
While it is not possible to have the 'old' memory leaks with smart pointer it
is possible to waste memory still. By adding a shared_ptr of this to the
lambda, that is added to the queue of this the smart pointer can't delete this
object.

The memory leak appears if the user frees the cache (free_all is called)