Fix include paths with spaces
If a space was previously in the include path found in the CMake file the path
would be split at the space in separate arguments that lead to ill formed
include arguments. Further did the clang_parseTranslationUnit not interpret
the paths correctly what clang_createTranslationUnitFromSourceFile does.

The regex does not work if a path contains a " but I hope no one uses something
like that